refactor to scripts/
