+++ scripts/generate/README.md
@@ -2,7 +2,7 @@
 
 The `test_generate` script provides a straightforward way to validate models, tokenizers, checkpoints, and device compatibility by running a single forward pass. This script functions as a sanity check to ensure everything is set up correctly.
 
-While **torchtitan** focuses on advanced features for distributed pre-training, this tool acts as a lightweight integration test to verify runtime setup. For more extensive inference and generation capabilities, consider tools like [pytorch/torchchat](https://github.com/pytorch/torchchat/).
+While **torchtitan** focuses on advanced features for distributed pre-training, this script acts as a lightweight integration test to verify runtime setup. For more extensive inference and generation capabilities, consider tools like [pytorch/torchchat](https://github.com/pytorch/torchchat/).
 
 ## Purpose and Use Case
 
@@ -19,19 +19,19 @@ This script is ideal for users who need to:
 ```bash
 NGPU=1 CONFIG_FILE=./train_configs/llama3_8b.toml CHECKPOINT_DIR=./outputs/checkpoint/ \
 PROMPT="What is the meaning of life?" \
-./test/generate/run_llama_pred.sh --max_new_tokens=32 --temperature=0.8 --seed=3
+./scripts/generate/run_llama_generate.sh --max_new_tokens=32 --temperature=0.8 --seed=3
 ```
 
-#### Run on 4 GPUs
+#### Run on 4 GPUs and pipe results to a json file.
 
 ```bash
 NGPU=4 CONFIG_FILE=./train_configs/llama3_8b.toml CHECKPOINT_DIR=./outputs/checkpoint/ \
 PROMPT="What is the meaning of life?" \
-./test/generate/run_llama_pred.sh --max_new_tokens=32 --temperature=0.8 --seed=3
+./scripts/generate/run_llama_generate.sh --max_new_tokens=32 --temperature=0.8 --seed=3 --out > output.json
 ```
 
 #### View Available Arguments
 
 ```bash
-> python ./test/generate/test_generate.py --help
+> python ./scripts/generate/test_generate.py --help
 ```
